Handle api error
