+++ ShiftsLogger.hasona23/ShiftsLogger.sln
