+++ ShiftsLogger.hasona23/ShiftsLogger/Models/Shift.cs
