+++ ShiftsLogger.hasona23/ShiftsLogger/Models/Worker.cs
