+++ ShiftsLogger.hasona23/ShiftsLogger/Program.cs
