+++ ShiftsLogger.hasona23/ShiftsLogger/Properties/launchSettings.json
