+++ ShiftsLogger.hasona23/ShiftsLogger/Services/WorkerService.cs
