+++ ShiftsLogger.hasona23/ShiftsLogger/ShiftsLoggerContext.cs
