+++ ShiftsLogger.hasona23/ShiftsLogger/appsettings.Development.json
