+++ ShiftsLogger.hasona23/ShiftsLogger/appsettings.json
@@ -8,7 +8,7 @@
   "AllowedHosts": "*",
   
     "ConnectionStrings": {
-      "DefaultConnection": "Server=localhost\\SQLEXPRESS;Database=ShiftsLoggerDB;User Id=hasona23ShiftsLogger;Password=pass1234;"
+      "DefaultConnection": "Server=(localdb)\\mssqllocaldb;Database=ShiftsLoggerDB;Trusted_Connection=True"
     }
   
 }
