+++ ShiftsLogger.hasona23/ShiftsLoggerCLI/Enums/Options.cs
