+++ ShiftsLogger.hasona23/ShiftsLoggerCLI/MenuBuilder.cs
