+++ ShiftsLogger.hasona23/ShiftsLoggerCLI/Models/Shift.cs
