+++ ShiftsLogger.hasona23/ShiftsLoggerCLI/Models/Worker.cs
