+++ ShiftsLogger.hasona23/ShiftsLoggerCLI/PositionManager.cs
