+++ ShiftsLogger.hasona23/ShiftsLoggerCLI/Program.cs
