+++ ShiftsLogger.hasona23/ShiftsLoggerCLI/ShiftsLoggerCLI.csproj
