+++ ShiftsLogger.hasona23/ShiftsLoggerCLI/VisualisationEngine.cs
